considering how to manage the operation class, which is getting quite large
