+++ src/app/core/model/drafts.ts
@@ -238,6 +238,32 @@ export const createDraft = (
 
 
   /**
+   * creates a new draft of the size specified using the data from the input draft. 
+   * since the input draft can be of a different size, this will resize data as neccessary.
+   * @param to_copy 
+   * @param warps 
+   * @param wefts 
+   * @returns a new draft object
+   */
+  export const generateDraftUsingAnotherDraft = (
+    to_copy: Draft,
+    warps: number,
+    wefts: number 
+  ) : Draft =>{
+
+    const draft = initDraft();
+    draft.drawdown = generateDrawdownWithPattern(to_copy.drawdown, warps, wefts);
+    draft.rowSystemMapping = generateMappingFromPattern(draft.drawdown, to_copy.rowSystemMapping, 'row');
+    draft.rowShuttleMapping = generateMappingFromPattern(draft.drawdown, to_copy.rowShuttleMapping, 'row');
+    draft.colSystemMapping = generateMappingFromPattern(draft.drawdown, to_copy.colSystemMapping, 'col');
+    draft.colShuttleMapping = generateMappingFromPattern(draft.drawdown, to_copy.colShuttleMapping, 'col');
+
+    return draft;
+
+  }
+
+
+  /**
    * creates a drawdown of a given size filled with the specified pattern
    * @param pattern the pattern to fill this drawdown with
    * @param width the width / number of warps in the result
@@ -266,6 +292,7 @@ export const createDraft = (
     
   
   }
+
 
 
   /**
